deleted meta tuple file
