+++ go/store/prolly/node_cursor.go
@@ -51,7 +51,7 @@ type searchFn func(item nodeItem, nd Node) (idx int)
 func newCursorAtStart(ctx context.Context, nrw NodeStore, nd Node) (cur *nodeCursor, err error) {
 	cur = &nodeCursor{nd: nd, nrw: nrw}
 	for !cur.isLeaf() {
-		nd, err = fetchChild(ctx, nrw, cur.currentRef())
+		nd, err = fetchChildNode(ctx, nrw, cur.currentRef())
 		if err != nil {
 			return nil, err
 		}
@@ -67,7 +67,7 @@ func newCursorPastEnd(ctx context.Context, nrw NodeStore, nd Node) (cur *nodeCur
 	cur.skipToNodeEnd()
 
 	for !cur.isLeaf() {
-		nd, err = fetchChild(ctx, nrw, cur.currentRef())
+		nd, err = fetchChildNode(ctx, nrw, cur.currentRef())
 		if err != nil {
 			return nil, err
 		}
@@ -102,7 +102,7 @@ func newCursorAtItem(ctx context.Context, nrw NodeStore, nd Node, item nodeItem,
 		// stay in bounds for internal nodes
 		cur.keepInBounds()
 
-		nd, err = fetchChild(ctx, nrw, cur.currentRef())
+		nd, err = fetchChildNode(ctx, nrw, cur.currentRef())
 		if err != nil {
 			return cur, err
 		}
@@ -126,7 +126,7 @@ func newLeafCursorAtItem(ctx context.Context, nrw NodeStore, nd Node, item nodeI
 		cur.keepInBounds()
 
 		// reuse |cur| object to keep stack alloc'd
-		cur.nd, err = fetchChild(ctx, nrw, cur.currentRef())
+		cur.nd, err = fetchChildNode(ctx, nrw, cur.currentRef())
 		if err != nil {
 			return cur, err
 		}
@@ -222,7 +222,7 @@ func (cur *nodeCursor) seek(ctx context.Context, item nodeItem, cb compareFn) (e
 		// stay in bounds for internal nodes
 		cur.parent.keepInBounds()
 
-		cur.nd, err = fetchChild(ctx, cur.nrw, cur.parent.currentRef())
+		cur.nd, err = fetchChildNode(ctx, cur.nrw, cur.parent.currentRef())
 		if err != nil {
 			return err
 		}
@@ -347,7 +347,7 @@ func (cur *nodeCursor) retreatInBounds(ctx context.Context) (bool, error) {
 // It's called whenever the cursor advances/retreats to a different chunk.
 func (cur *nodeCursor) fetchNode(ctx context.Context) (err error) {
 	assertTrue(cur.parent != nil)
-	cur.nd, err = fetchChild(ctx, cur.nrw, cur.parent.currentRef())
+	cur.nd, err = fetchChildNode(ctx, cur.nrw, cur.parent.currentRef())
 	cur.idx = -1 // caller must set
 	return err
 }
@@ -388,6 +388,11 @@ func (cur *nodeCursor) copy(other *nodeCursor) {
 	} else {
 		assertTrue(other.parent == nil)
 	}
+}
+
+func fetchChildNode(ctx context.Context, ns NodeStore, ref hash.Hash) (Node, error) {
+	// todo(andy) handle nil Node, dangling ref
+	return ns.Read(ctx, ref)
 }
 
 func assertTrue(b bool) {
